add new classes and configuration with spring and annotations

+++ src/main/java/edu/unq/desapp/grupo_a/backend/model/User.java
@@ -5,21 +5,15 @@ import java.util.Collection;
 
 import edu.unq.desapp.grupo_a.backend.model.exceptions.UserInitException;
 
-import javax.persistence.Access;
-import javax.persistence.AccessType;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import javax.persistence.OneToOne;
 import javax.persistence.Table;
-import javax.persistence.Transient;
 
 import org.hibernate.annotations.LazyCollection;
 import org.hibernate.annotations.LazyCollectionOption;
@@ -63,7 +57,7 @@ public class User extends PersistenceEntity{
 		this.name = name;
 		this.address = address;
 		this.email = email;
-		this.reputation = (double) 0;
+		this.reputation = 0;
 		this.vehicles = new ArrayList<>();
 		this.creditAccount = new CreditAccount();
 	}
@@ -115,9 +109,7 @@ public class User extends PersistenceEntity{
 		this.reputation = reputation;
 	}
 
-	//@Access(AccessType.PROPERTY)
 	@OneToOne(cascade = CascadeType.ALL, fetch = FetchType.EAGER)
-	//@JoinColumn(name="user_id")
 	public CreditAccount getCreditAccount() {
 		return creditAccount;
 	}
